Cross-product automata + use KdTree for state search

+++ generation/src/cmd.rs
@@ -38,7 +38,7 @@ pub enum Command {
         #[arg(short, long, default_value_t=false, help="Add noise in the output file")]
         noise: bool,
         #[arg(short, long, default_value_t=10, help="Minimum number of flows to generate.")] // TODO: remove default value for release
-        flow_count: i32,
+        flow_count: u64,
         #[arg(short='d', long, default_value=None, help="Unix time for the beginning of the pcap. By default, use current time.")]
         start_unix_time: Option<u64>,
         #[arg(short, long, help="Seed for random number generation")]
